Added encoding
Added functions to encode the input data into a vector to be further used for semantics (one more manual and ceremonial, the other leveraging pre-built functions more and to be used in the future. A proper value of the char-to-token ratio is estimated using a custom function to help in efficient chunking based on the available token limit of the chosen model.